src/modules/OCR/requirements.txt
#! Python3.7

paddlepaddle>=2.6.0             # Installing PaddelPaddle, the Deep Learning platform
paddleocr==2.7.0.3              # Installing PaddleOCR, the OCR toolkit based on PaddlePaddle

imutils                         # Installing imutils, the image utilities library
Pillow<10.0.0                   # Installing Pillow, a Python Image Library
opencv-python                   # Installing OpenCV, the Computer Vision library for Python
numpy                           # Installing NumPy, a package for scientific computing

# end of file